Added db2_exec_nofail and support for admin nodes
When a newly created instance is configured, puppet fails when
determining whether or not nodes, dcs and database entries exist because
the command "db2 LIST * DIRECTORY" will cause db2 to throw an error
because the directory doesn't exist.  There are various scenarios here
so the best way is to allow db2_exec to fail when being called from the
exists? method.  Therefore all exists? methods now call db2_exec_nofail

Previously, any nodes added as admin => true would cause puppet to try
and configure them again on the next run, that's because we need to
separate commands to list the configured nodes, list node directory and
list admin node directory.  This commit adds support to the
db2_catalog_node provider to query both types of node entry and return a
unified result.